Move bin-bound scoring tests to a dedicated test file
CodeScene flagged the mirror test module as losing cohesion purely from the
added #549/#550 regression tests. Relocate them to a small, internally
cohesive tests/evaluate/test_uncertainty_metrics_bounds.py so the existing
module is no longer grown, leaving its cohesion unchanged.
